refactor: rename configuration files and simplify defaults
Rename configuration files to improve clarity. The default configurations for Logger, Meter, and Tracer are now initialized using dedicated functions instead of inline definitions. This change enhances maintainability by centralizing default configurations. The meter options and registration functionalities have been removed as they are no longer necessary with the new structure.

Breaking changes:
- Renamed config/parse.go to config/default.go
- Renamed meter/constants/exporter.go to meter/global.go
- Renamed meter/config.go to meter/meter.go
- Removed meter/options.go and meter/register.go
